Move keystore to legacy directory.
