+++ legacy/keystore/keystore_test.go
